+++ src/pygsdata/plots.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import matplotlib.pyplot as plt
 import numpy as np
 
-from .pygsdata import GSData
+from .gsdata import GSData
 
 
 def plot_waterfall(
